+++ src/GUI/QuestMarker.cpp
@@ -4,13 +4,13 @@
 #include "CharacterCore.h"
 #include "GameObjectComponents/TooltipComponent.h"
 
-const float QuestMarker::SIZE = 16.f;
+const int QuestMarker::SIZE = 16;
 
 QuestMarker::QuestMarker(const QuestData& questData, const CharacterCore* core) {
 	m_questData = questData;
 	m_characterCore = core;
 	
-	setBoundingBox(sf::FloatRect(0.f, 0.f, SIZE, SIZE));
+	setBoundingBox(sf::FloatRect(0.f, 0.f, static_cast<float>(SIZE), static_cast<float>(SIZE)));
 	m_sprite.setTexture(*g_resourceManager->getTexture(GlobalResource::TEX_GUI_QUESTMARKERS));
 
 	m_tooltipComponent = new TooltipComponent("", this);
